--- frontend/src/components/dump.vue
@@ -1,482 +1,0 @@
-<template>
-  <div class="m-doc">
-    <div class="m-header">
-      <div class="f-tit">
-        <h3 class="f-fl">FuckDb</h3>
-        <el-dropdown class="f-fl m-dropdown" @command="handleCommand">
-          <span class="el-dropdown-link">
-            Navigation
-            <i class="el-icon-arrow-down el-icon--right"></i>
-          </span>
-          <el-dropdown-menu slot="dropdown">
-            <el-dropdown-item command="FuckDb" icon="el-icon-price-tag">
-              <router-link :to="{path:'/'}">FuckDb</router-link>
-            </el-dropdown-item>
-            <el-dropdown-item command="dump" icon="el-icon-view">
-              <router-link :to="{path:'dump'}">dump</router-link>
-            </el-dropdown-item>
-            <el-dropdown-item command="json-to-go" icon="el-icon-s-tools">json-to-go</el-dropdown-item>
-            <el-dropdown-item command="yaml-to-json" icon="el-icon-s-tools">yaml-to-json</el-dropdown-item>
-            <!-- <el-dropdown-item command="struct-to-sql">struct-to-sql</el-dropdown-item> -->
-          </el-dropdown-menu>
-        </el-dropdown>
-        <iframe
-          style="margin-left: 15px; margin-bottom:-7px;"
-          frameborder="0"
-          scrolling="0"
-          width="91px"
-          height="20px"
-          src="https://ghbtns.com/github-btn.html?user=hantmac&repo=fuckdb&type=star&count=true"
-        ></iframe>
-      </div>
-    </div>
-
-    <div class="m-fuckDb" v-show="tabName == 'FuckDb'">
-      <el-form ref="form" :model="form" :rules="rules" label-width="130px" size="medium">
-        <el-row>
-          <el-col :span="24">
-            <el-form-item label="Change DB">
-              <el-select
-                v-model="value"
-                placeholder="select DB"
-                @change="changeDBList"
-                clearable
-                @clear="clearDBList"
-              >
-                <el-option
-                  v-for="item in optionsDBLog"
-                  :key="item.value"
-                  :label="item.label"
-                  :value="item.value"
-                ></el-option>
-              </el-select>
-              <el-button @click="cleanDb">Clear</el-button>
-            </el-form-item>
-          </el-col>
-        </el-row>
-        <el-row>
-          <el-col :span="8">
-            <el-form-item prop="mysql_host" label="mysql_host">
-              <el-input v-model="form.mysql_host" clearable></el-input>
-            </el-form-item>
-          </el-col>
-
-          <el-col :span="8">
-            <el-form-item prop="mysql_port" label="mysql_port">
-              <el-input v-model.number="form.mysql_port" clearable></el-input>
-            </el-form-item>
-          </el-col>
-
-          <el-col :span="8">
-            <el-form-item prop="mysql_db" label="mysql_db">
-              <el-input v-model="form.mysql_db" clearable></el-input>
-            </el-form-item>
-          </el-col>
-        </el-row>
-
-        <el-row>
-          <el-col :span="8">
-            <el-form-item prop="mysql_table" label="mysql_table">
-              <el-input v-model="form.mysql_table" clearable></el-input>
-            </el-form-item>
-          </el-col>
-
-          <el-col :span="8">
-            <el-form-item prop="mysql_passwd" label="mysql_passwd">
-              <el-input type="password" v-model="form.mysql_passwd" clearable></el-input>
-            </el-form-item>
-          </el-col>
-
-          <el-col :span="8">
-            <el-form-item prop="mysql_user" label="mysql_user">
-              <el-input v-model="form.mysql_user" clearable></el-input>
-            </el-form-item>
-          </el-col>
-        </el-row>
-
-        <el-row>
-          <el-col :span="8">
-            <el-form-item label="请选择输出格式">
-              <el-select v-model="form.view_type" placeholder="请选择">
-                <el-option
-                  v-for="item in viewOptions"
-                  :key="item.value"
-                  :label="item.label"
-                  :value="item.value">
-                </el-option>
-              </el-select>
-          </el-col>
-        </el-row>
-
-        <el-form-item>
-          <div class="f-fl">
-            <el-button type="primary" @click="onSubmit('form')">Create</el-button>
-            <el-button @click="resetForm('form')">Reset</el-button>
-          </div>
-          <div class="u-tips">
-            <p>
-              <b>声明：</b> 本工具不会记录任何用户信息，请放心使用
-            </p>
-            <p>
-              <b>Disclaimer:</b> This tool does not record any user information, please use it with confidence
-            </p>
-          </div>
-        </el-form-item>
-      </el-form>
-      <div class="m-body">
-        <highlight-code lang="golang">{{code}}</highlight-code>
-      </div>
-    </div>
-
-    <div class="m-tool" v-show="tabName == 'json-to-go'">
-      <iframe src="https://mholt.github.io/json-to-go/" frameborder="0"></iframe>
-    </div>
-    <div class="m-tool" v-show="tabName == 'yaml-to-json'">
-      <iframe src="https://onlineyamltools.com/convert-yaml-to-json" frameborder="0"></iframe>
-    </div>
-    <div class v-show="tabName == 'struct-to-sql'">
-      <div class="m-tabs">
-        <el-button-group>
-          <el-button size="medium" icon="el-icon-plus">Create</el-button>
-          <el-button size="medium" icon="el-icon-delete">Clean</el-button>
-          <el-button size="medium" icon="el-icon-copy-document">Copy</el-button>
-        </el-button-group>
-      </div>
-      <editor
-        class="m-editor"
-        height="400px"
-        width="100%"
-        ref="editor"
-        :content="content"
-        :options="{
-        enableBasicAutocompletion: true,
-        enableSnippets: true,
-        enableLiveAutocompletion: true,
-        tabSize:2
-    }"
-        :fontSize="12"
-        :lang="'golang'"
-        :theme="'textmate'"
-        @onChange="editorChange"
-        @init="editorInit"
-      ></editor>
-      <highlight-code lang="sql">{{codeSql}}</highlight-code>
-    </div>
-  </div>
-</template>
-
-<script>
-import service from "../config/api";
-import API from "../config/index";
-import Editor from "vue2x-ace-editor";
-
-export default {
-  inject:['reload'],
-  name: "dump",
-  components: {
-    Editor
-  },
-  data() {
-    return {
-      code: "package test",
-      content: `
-//code...
-type testDB struct {
-    Status        string \`gorm:"column:Status" json:"Status" xml:"Status"\`
-}
-`,
-      tabName: "FuckDb", //"struct-to-sql"
-      codeSql: `
-      CREATE TABLE \`auto_generated\`
-      (
-        \`registration_date\` varchar(128) NOT NULL ,
-
-      ) engine=innodb DEFAULT charset=utf8mb4;
-      `,
-      form: {
-        mysql_host: "",
-        mysql_port: 3306,
-        mysql_db: "",
-        mysql_table: "",
-        mysql_passwd: "",
-        mysql_user: "",
-        view_type:"md",
-      },
-      rules: {
-        mysql_host: [
-          { required: true, message: "mysql_host is empty", trigger: "blur" }
-        ],
-        mysql_db: [
-          { required: true, message: "mysql_db is empty", trigger: "blur" }
-        ],
-        mysql_table: [
-          { required: true, message: "mysql_table is empty", trigger: "blur" }
-        ],
-        mysql_passwd: [
-          { required: true, message: "mysql_passwd is empty", trigger: "blur" }
-        ],
-        mysql_user: [
-          { required: true, message: "mysql_user is empty", trigger: "blur" }
-        ],
-        view_type: [
-          { required: true, message: "view_type is empty", trigger: "blur" }
-        ],
-        mysql_port: [
-          {
-            required: true,
-            type: "number",
-            message: "port must be integer",
-            trigger: "blur"
-          }
-        ]
-      },
-      optionsDBLog: [],
-      FuckDbList: [],
-      viewOptions: [
-        {
-          value: 'md',
-          label: 'markdown'
-        }, {
-          value: 'json',
-          label: 'json'
-        }, {
-          value: 'txt',
-          label: 'txt'
-        },{
-          value: 'csv',
-          label: 'csv'
-        }
-      ],
-      value: ""
-    };
-  },
-  mounted() {
-    this.setFuckDbChangeDBList();
-  },
-  methods: {
-    editorInit() {
-      require("brace/ext/language_tools");
-      require(`brace/mode/golang`);
-      require(`brace/snippets/golang`);
-      require(`brace/theme/textmate`);
-    },
-    editorChange() {
-      this.$refs.editor.setCompleteData([
-        {
-          name: "test", //名称
-          value: "test", //值
-          caption: "test", //字幕，展示在列表的内容
-          meta: "test", //展示类型
-          type: "local", //类型
-          score: 1000 // 分数越大排在越上面
-        }
-      ]);
-    },
-    GetDb2view() {
-      let data = {
-        mysql_host: this.form.mysql_host,
-        mysql_port: this.form.mysql_port,
-        mysql_db: this.form.mysql_db,
-        mysql_table: this.form.mysql_table,
-        mysql_passwd: this.form.mysql_passwd,
-        mysql_user: this.form.mysql_user,
-        view_type:this.form.view_type
-      };
-
-      service({
-        url: `${API.APIdb2struct}/api/db2view`,
-        method: "post",
-        data
-      })
-        .then(res => {
-          if (res.data.status === "0") {
-            this.code = res.data.data;
-            this.setFuckDbList(this.form);
-          }
-        })
-        .catch(error => {
-          if (error.response.status == 500) {
-            try {
-              this.$message.error(error.response.data.error);
-            } catch (error) {
-              this.$message.error("异常错误！");
-            }
-          } else {
-            this.$message.error("异常错误！");
-          }
-        });
-    },
-    setFuckDbList(data) {
-      let FuckDbList = window.localStorage.getItem("FuckDb_dump");
-      if (FuckDbList) {
-        FuckDbList = JSON.parse(FuckDbList);
-        let FuckDbListDistinguish = false;
-
-        FuckDbList.forEach(obj => {
-          if (
-            obj.mysql_user === data.mysql_user &&
-            obj.mysql_host === data.mysql_host
-          ) {
-            FuckDbListDistinguish = true;
-            obj.mysql_db = data.mysql_db;
-            obj.mysql_table = data.mysql_table;
-          }
-        });
-
-        if (!FuckDbListDistinguish) {
-          FuckDbList.push(data);
-        }
-        window.localStorage.setItem("FuckDb_dump", JSON.stringify(FuckDbList));
-      } else {
-        FuckDbList = [data];
-        window.localStorage.setItem("FuckDb_dump", JSON.stringify([data]));
-      }
-      this.optionsDBLog = [];
-      FuckDbList.forEach((obj, index) => {
-        this.optionsDBLog.push({
-          value: index,
-          label: `${obj.mysql_user}@${obj.mysql_host}`
-        });
-        if (
-          obj.mysql_user === data.mysql_user &&
-          obj.mysql_host === data.mysql_host
-        ) {
-          this.value = index;
-        }
-      });
-      this.FuckDbList = FuckDbList;
-    },
-    setFuckDbChangeDBList() {
-      let FuckDbList = window.localStorage.getItem("FuckDb_dump");
-      if (FuckDbList) {
-        FuckDbList = JSON.parse(FuckDbList);
-        this.FuckDbList = FuckDbList;
-        this.optionsDBLog = [];
-        FuckDbList.forEach((obj, index) => {
-          this.optionsDBLog.push({
-            value: index,
-            label: `${obj.mysql_user}@${obj.mysql_host}`
-          });
-        });
-      }
-    },
-    changeDBList(index) {
-      if (index !== "") {
-        this.form = this.FuckDbList[index];
-      }
-    },
-    clearDBList() {
-      this.code = "package test";
-      this.form = {
-        mysql_host: "",
-        mysql_port: 3306,
-        mysql_db: "",
-        mysql_table: "",
-        mysql_passwd: "",
-        mysql_user: "",
-        view_type:""
-      };
-    },
-    cleanDb() {
-      this.value = "";
-      this.FuckDbList = [];
-      this.optionsDBLog = [];
-      this.code = "package test";
-      this.form = {
-        mysql_host: "",
-        mysql_port: 3306,
-        mysql_db: "",
-        mysql_table: "",
-        mysql_passwd: "",
-        mysql_user: "",
-        view_type:""
-      };
-
-      window.localStorage.setItem("FuckDb_dump", "");
-    },
-    onSubmit(formName) {
-      this.$refs[formName].validate(valid => {
-        if (valid) {
-          this.GetDb2view();
-        } else {
-          console.log("error submit!!");
-          return false;
-        }
-      });
-    },
-    resetForm(formName) {
-      this.value = "";
-      this.code = "package test";
-      this.$refs[formName].resetFields();
-      this.setFuckDbChangeDBList();
-    },
-    handleCommand(command) {
-      if (command === "FuckDb" || command==="dump") {
-        this.reload()
-        document.getElementById("app").style.overflowY = "auto";
-      } else {
-        document.getElementById("app").style.overflowY = "hidden";
-      }
-      this.tabName = command;
-    }
-  }
-};
-</script>
-
-<!-- Add "scoped" attribute to limit CSS to this component only -->
-<style scoped scss>
-h1,
-h2 {
-  font-weight: normal;
-}
-ul {
-  list-style-type: none;
-  padding: 0;
-}
-
-.m-doc {
-  position: relative;
-  height: 100%;
-}
-
-.m-header {
-  overflow: hidden;
-  padding: 10px 0;
-}
-.m-tool {
-  position: relative;
-}
-.f-tit {
-  margin-left: 10px;
-}
-.f-fl {
-  float: left;
-}
-.m-dropdown {
-  cursor: pointer;
-  margin: 5px 0 0 15px;
-  cursor: pointer;
-  color: #409eff;
-}
-.m-tool,
-.m-tool iframe {
-  width: 100%;
-  height: 100%;
-}
-.m-etitor-tool {
-  height: 39px;
-}
-.u-tips {
-  padding-left: 20px;
-  color: #aaa;
-  font-size: 12px;
-  float: left;
-  line-height: 18px;
-}
-.m-tabs {
-  padding: 10px 0 10px 20px;
-  background: #f5f7fa;
-}
-.m-editor {
-}
-</style>
