+++ frontend/src/router/index.js
@@ -1,7 +1,6 @@
 import Vue from 'vue'
 import Router from 'vue-router'
 const home = () => import('@/components/home')
-const dump = () => import('@/components/dump')
 
 Vue.use(Router)
 
@@ -11,13 +10,6 @@ export default new Router({
       path: '/',
       name: 'home',
       component: home
-    },{
-      path: '/dump',
-      name: 'dump',
-      component: dump
-    },{
-      path:'*',
-      redirect:'/'
     }
   ]
 })
